# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
    type: object
  api.SessionCreateRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securepassword
        type: string
    type: object
  api.UserCreateRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securepassword
        type: string
    type: object
  api.UserResponse:
    properties:
      email:
        example: user@example.com
        type: string
      id:
        example: 1
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://example.com/support
  description: API with authentication and user management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Auth API
  version: "1.0"
paths:
  /users/me:
    get:
      consumes:
      - application/json
      description: Returns information about the current authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Information about the current authenticated user
      tags:
      - auth
  /users/signin:
    post:
      consumes:
      - application/json
      description: Creates a session for the user
      parameters:
      - description: Pass data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/api.SessionCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Enter the system
      tags:
      - auth
  /users/signup:
    post:
      consumes:
      - application/json
      description: Creates a new user in the system
      parameters:
      - description: User data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/api.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create a new user
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"